recommendation 수정

+++ src/main/java/com/example/demo/user/repository/AnimeRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 
 public interface AnimeRepository extends JpaRepository<Anime, Long> {
     boolean existsByTitle(String title);
-    Optional<Anime> findByMalId(Integer malId);
+    Optional<Anime> findByMalId(Long malId);
     Page<Anime> findByTitleContaining(String keyword, Pageable pageable);
     boolean existsByMalId(Integer malId);
     @Query("select a.malId from Anime a")
